Move more leaf functions to bukuutil (which will be cleaned up later)
